cost-bot/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.23.1
requests==2.31.0
#azure-ai-openai==1.1.0
python-dotenv==1.0.1
openai>=1.40.0
#azure-ai-openai==1.0.0b6

openai>=1.30.0
azure-search-documents>=11.4.0
azure-core>=1.30.0